+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.12.1 [unreleased]
 
 ### Features
+- [#976](https://github.com/influxdata/telegraf/pull/976): Reduce allocations in the UDP and statsd inputs.
 
 ### Bugfixes
 - [#968](https://github.com/influxdata/telegraf/issues/968): Processes plugin gets unknown state when spaces are in (command name)